AVRO-2921: Add Type Hints for avro.name
Refactor name registry into avro.name and add type hints.
